fix: Code refactoring. Introduce includeProtos
