+++ docs/mocking-gRPC.md
@@ -121,4 +121,22 @@ module.exports.plconfig = {
 }
 ```
 
-Full list of the available options can be found in [@grpc/proto-loader](https://www.npmjs.com/package/@grpc/proto-loader){target=\_blank} README.
+Available options are as follows:
+
+| Field name | Valid values | Description
+|------------|--------------|------------
+| `keepCase` | `true` or `false` | Preserve field names. The default is to change them to camel case.
+| `longs` | `String` or `Number` | The type to use to represent `long` values. Defaults to a `Long` object type.
+| `enums` | `String` | The type to use to represent `enum` values. Defaults to the numeric value.
+| `bytes` | `Array` or `String` | The type to use to represent `bytes` values. Defaults to `Buffer`.
+| `defaults` | `true` or `false` | Set default values on output objects. Defaults to `false`.
+| `arrays` | `true` or `false` | Set empty arrays for missing array values even if `defaults` is `false` Defaults to `false`.
+| `objects` | `true` or `false` | Set empty objects for missing object values even if `defaults` is `false` Defaults to `false`.
+| `oneofs` | `true` or `false` | Set virtual oneof properties to the present field's name. Defaults to `false`.
+| `json` | `true` or `false` | Represent `Infinity` and `NaN` as strings in `float` fields, and automatically decode `google.protobuf.Any` values. Defaults to `false`
+| `includeDirs` | An array of strings | A list of search paths for imported `.proto` files.
+| `includeProtos` | An array of strings | A list of proto files to be loaded. If specified, Camouflage will only load the specified proto files and ignore other protofiles in `config.protocols.grpc.protos_dir`.
+
+!!!note
+
+    Camouflage extends the configurations provided by Protoloader Options([@grpc/proto-loader](https://www.npmjs.com/package/@grpc/proto-loader){target=\_blank}).
